pyproject.toml
[project]
name = "budget-app-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.3",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",
    "whitenoise[brotli]>=6.8.2",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "django-redis>=5.4.0",
    "psycopg2>=2.9.10",
    "pillow>=11.0.0",
    "argon2-cffi>=23.1.0",
    "hiredis>=3.0.0",
    "flower>=2.0.1",
    "djangorestframework-simplejwt>=5.3.1",
    "dj-rest-auth[with-social]>=7.0.0",
    "django-money>=3.5.3",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4.6",
    "ruff>=0.8.0",
    "watchdog>=6.0.0",
    "watchfiles>=0.24.0",
    "werkzeug[watchdog]>=3.1.3",
]

[tool.ruff]
line-length = 88 
target-version = "py312"

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "D",   # pydocstyle
    "DJ",  # flake8-django
    "INP", # flake8-no-pep420
]
ignore = [
    "D100",
    "D104",
    "D105",
    "D106",
    "D107",
]
fixable = ["D200", "E401", "I"]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
